[FE] refactor: #72 #74 CalendarContainer 리팩토링
Container로 기존 구현했던 내용을 옮기고 일부 리팩토링 진행했습니다.
